+++ index.js
@@ -1,8 +1,0 @@
-'use strict';
-const kit = module.exports;
-
-// fs-extra
-kit.fs = require('./fs');
-
-// superagent
-kit.request = require('./request');