+++ CHANGELOG.md
@@ -1,17 +1,27 @@
 # Changelog
 
+## 2024-03-06 v3.0.0
+
+- rm previous content
+- add `lnsfSafe` symlink function
+
 ## 2016-08-30 v2.0.0
+
 - rm `req.endAsync`
 
 ## 2016-03-09 v1.2.0
+
 - update `promise.ify`, include bug fix
 
 ## 2016-03-09 v1.1.0
+
 - instead `bluebird` with native Promise, and use `promise.ify`
 - remove co
 
 ## unknown date v1.0.0
+
 - it's 1.0
 
 ## 2016-03-06 v0.1.0
+
 - 其他都使用这个了